Breaking out logic, so its easier to test
